Add C++ YieldChannel class to read nucleosynthetic yield data
Adds src/yields, a new namespace mirroring src/tracks's own structure:

- YieldCommons.hpp declares Channel (currently ccsn_ and
  massiveStarWinds_) and the default yields.toml registry path.
- YieldChannel reads one channel's yield data for one model over a
  requested [Fe/H] range, following the same registry-then-HDF5
  loading pattern as Tracks3D/Extinct: validates the channel/model
  against yields.toml, then treats the HDF5 file's own feh_<value>
  group attributes as authoritative for which [Fe/H] values actually
  exist (mirroring findMatchingTracks()'s identical "scan groups by
  attribute, don't trust the registry's own Fe_H list" approach),
  bracketing [fehMin, fehMax] with the minimal enclosing subset and
  throwing if the request can't be accommodated at all -- mirroring
  Tracks3D's own up-front range check.

yld() builds its mdspan view fresh from yieldData_ on every call
rather than caching it as its own sibling member: mirroring
Mesh3DInterpolator's identical x()/y()/z()/f() accessors, this is
what keeps copying or moving a YieldChannel safe (a cached view would
keep pointing at another object's own, separately-relocated backing
storage) with a plain, compiler-generated copy/move -- verified by a
dedicated test that reads back the correct value from a copied and a
move-constructed YieldChannel alike.

Also fixes a data-format issue this exercise surfaced before it went
into a real PR: import_yield_tables.py wrote isotope_z/isotope_a as
int32, which had no existing HDF5Utils reader; switched to float64,
matching data/elem/isotopes.h5's own Z/A datasets, so YieldChannel
reads them with the same utils::readDataset1D() every other float
dataset already goes through.

tests/yields/assets/yields_test.h5 (via
data/tools/yields/make_yields_test_fixture.py) extracts 2 masses and
3 isotopes from the real data/yields/sukhbold16.h5, so
testYieldChannel.hpp's assertions check against real Sukhbold et al.
(2016) values already independently verified by hand against the
original text files, including the zero-filled ccsn yield for a
failed-supernova mass.

+++ data/tools/yields/import_yield_tables.py
@@ -214,8 +214,14 @@ def import_directory(input_dir: str) -> dict:
 
     masses = sorted(tables_by_mass.keys())
     isotopes = sorted(all_isotopes, key=isotope_z_a)
-    isotope_z = np.array([isotope_z_a(iso)[0] for iso in isotopes], dtype=np.int32)
-    isotope_a = np.array([isotope_z_a(iso)[1] for iso in isotopes], dtype=np.int32)
+    # float64, not an integer dtype -- matching data/elem/isotopes.h5's
+    # own Z/A datasets (see data/tools/elem/build_isotope_table.py), so
+    # the C++ reader can use the same utils::readDataset1D() every
+    # other float dataset in this project's own HDF5 files already
+    # goes through, rather than needing a separate integer-dataset
+    # reader just for this one case.
+    isotope_z = np.array([isotope_z_a(iso)[0] for iso in isotopes], dtype=np.float64)
+    isotope_a = np.array([isotope_z_a(iso)[1] for iso in isotopes], dtype=np.float64)
     isotope_index = {iso: i for i, iso in enumerate(isotopes)}
     mass_index = {m: j for j, m in enumerate(masses)}
